requirements.txt
dlib==19.24.2
lmdb==1.4.1
numpy==1.24.3
opencv_python==4.7.0.72
Pillow==10.1.0
PyYAML==6.0.1
Requests==2.31.0
scipy==1.9.1
timm==0.9.2
torch==1.7.1
torchvision==0.8.2
tqdm==4.65.0
wandb==0.15.5
kornia==0.7.0
